+++ src/SmartFormat/Extensions/DictionarySource.cs
@@ -17,14 +17,6 @@ namespace SmartFormat.Extensions
     /// </summary>
     public class DictionarySource : Source
     {
-        /// <summary>
-        /// CTOR.
-        /// </summary>
-        /// <param name="formatter"></param>
-        public DictionarySource(SmartFormatter formatter) : base(formatter)
-        {
-        }
-
         /// <inheritdoc />>
         public override bool TryEvaluateSelector(ISelectorInfo selectorInfo)
         {